[FEATURE] Debugging proxy
The debugging proxy is intended to sit between a debugger engine and an
IDE to transmit packages back and forth, with the possiblity to
intercept them. This is useful e.g. to learn about what is going on
during a debugging session or to change data on-the-fly, e.g. for
rewriting paths in breakpoints (if that is not supported by the IDE or
necessary for other reasons).

Currently, the proxy only dumps everything that comes through to the
terminal. Further possibilites, e.g. listeners to handle the data, can
be added in future commits.