[FEATURE] React to incoming debugger connections
The debug proxy had a working, but fully hardcoded approach to handling
the sessions. As this is intended to be a reusable library, we gradually
switch to using events for passing data around. This commit is the first
for this and adds the generic stream handling infrastructure plus a
rough event handling part. Currently, there is no further action taken
when a debugger session is opened.