[TASK] Remove event listening from message parser
As the stream knows about the parser because it is its sink, we don’t
need the additional connection via events currently.

In the future it might be necessary to switch the whole incoming data
processing to events (if we want to have further processing
capabilities), but currently this simple direct solution is enough.