[FEATURE] Transactions and result processing
This adds transactions and command result processing. The transaction is
created as the command is sent, and the result will later on be added
when it was returned by the debugger engine (this is an asynchronous
process).

Currently, the only fully implemented command is "run", which starts or
continues program execution. The result of that will be a status report;
if a breakpoint is set, it will most likely report a hit for that
breakpoint, otherwise (for small programs) the programm will have ended
already when the run command was processed.

From this state on, it is easily possible to implement further commands;
for some commands, it might be desirable to have further postprocessing
options (e.g. a method that is invoked after the response was fetched).