[FEATURE] Fetch property values from debugger engine
This adds support for the property_get DBGp command. As the process is
asynchronous, the value can only be processed in a callback. This is
done with "promises", using the implementation by React (reactphp.org).

To fetch a variable, simply instantiate a PropertyGet command, get its
promise and assign a callback:

```
$command = new PropertyGet($event->getSession(), '$myVariable');
$command->promise()->then(function($value) {
  echo "Fetched variable \$myVariable: $value\n";
});
```