[FEATURE] Disable autorun in debug sessions
When further actions should be triggered for a new debug session, the
session should only be started after these operations have finished. For
synchronous operations, this is no problem, but e.g. asynchronous calls
to the debugger engine pose a problem.

This is mitigated with this commit by making the autorun optional. To
disable autorun, call `disableAutorun()` on the debug session object.
Keep in mind that you have to call `run()` when the session is
completely initialized.