[BUGFIX] Decoding fails if packet length at end of window
If a packet ends in the middle of a packet length (e.g. 2|02), this was
not detected. Therefore, parsing the new message failed.

This is now fixed by first reading all contents and then decoding the
data as one large string. We assume that the data is all ready at the
socket, otherwise decoding might still fail. A fix for this is however
not so easy, therefore we don’t implement it with this commit.