[FEATURE] Evaluate expressions in debugger engine
This implements support for DBGp’s "eval" command. The command class
has to be named "Evaluate", as apparently the Composer autoloader does
not support underscores in _ (otherwise I’d have named it Eval_, as the
PhpParser package does it; this however resulted in "class not found"
errors in the tests).