[BUGFIX] Breakpoint hit should not continue automatically
When a breakpoint is hit, we have a default handler with priority -1000
that continues execution (i.e. sends a run command to the debugger
engine). This however makes it impossible to wait for answers from the
server, which will almost certainly be a problem when fetching data for
analysis. Therefore, the automatic continuation is only enabled when
autorun is also enabled for the session (which is the default; it can be
disabled by calling `disableAutorun()`.